Oops more updates

+++ ptah/cli/__init__.py
@@ -110,15 +110,17 @@ def dashboard():
 
 
 @app.command()
-def nuke():
+def nuke(docker: bool = True, kind: bool = True):
     """
     Forcibly delete the Kind cluster, all related resources, and prune dangling Docker images.
     """
     _forward(kill=True)
 
-    get(Docker).prune()
+    if docker:
+        get(Docker).prune()
 
-    get(Kind).delete()
+    if kind:
+        get(Kind).delete()
 
     filesystem = get(Filesystem)
     filesystem.delete(filesystem.cache_location())
